+++ app/src/routes/home/index.js
@@ -5,7 +5,7 @@ const router = express.Router();
 
 // 컨트롤러
 const ctrl = require("./home.ctrl");
-const webtoonctrl = require("../webtoon/webtoon.ctrl");
+const webtoonctrl = require("../webtoon/webtooncontroller");
 
 router.get("/", ctrl.output.home);
 router.get("/api/webtoons", webtoonctrl.process.getWebtoons);
